@@ -1,22 +1,20 @@
 name: flutter_midi_command
 description: A Flutter plugin for sending and receiving MIDI messages between Flutter and physical and virtual MIDI devices. Wraps CoreMIDI and android.media.midi in a thin dart/flutter layer.
-version: 0.4.9
+version: 0.4.15
 homepage: https://github.com/InvisibleWrench/FlutterMidiCommand
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.15.0 <4.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_midi_command_platform_interface: ^0.4.0
-    #path: ../flutter_midi_command_platform_interface
-
-#  flutter_midi_command_linux: ^0.1.5
+  flutter_midi_command_platform_interface: ^0.4.1
+  flutter_midi_command_linux: ^0.3.0
 
 dev_dependencies:
-  flutter_test: 
+  flutter_test:
     sdk: flutter
 
 # For information on the generic Dart part of this file, see the
